pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "perfectconfig"
version = "1.1.3"
dependencies = [
    "appdirs",
    "pyyaml",
    "werkzeug"
]
requires-python = ">=3.11"
authors = [
  {name = "Supratik Chatterjee", email = "supratikdevm96@gmail.com"},
]
maintainers = [
  {name = "Supratik Chatterjee", email = "supratikdevm96@gmail.com"},
]
description = "Configuration Managment library"
readme = "README.md"
license = {text = "MIT License"}
keywords = ["configuration", "management", "config"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Software Development",
  "Topic :: System :: Filesystems"
]

[project.urls]
Homepage = "https://conceivilize.com/products/perfectconfig"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/supratikchatterjee16/perfectconfig.git"
"Bug Tracker" = "https://github.com/supratikchatterjee16/perfectconfig/issues"
Changelog = "https://github.com/supratikchatterjee16/iam/blob/master/CHANGELOG.md"
